+++ Euler/Problems/EulerProblem10.cs
@@ -6,8 +6,8 @@ namespace Euler.Problems {
 
 	public class EulerProblem10 : Problem {
 		public EulerProblem10()
-			: base((long)10, 17, (long)2000000) {
-				SolutionResponse = 142913828922;
+      : base((long)10, (long)17, (long)2000000) {
+        SolutionResponse = (long)142913828922;
 		}
 
 		public override object Run(RunModes runMode, object input, bool Logging) {
