+++ Euler/Problems/EulerProblem3.cs
@@ -10,7 +10,7 @@ namespace Euler.Problems {
   public class EulerProblem3 : Problem {
     public EulerProblem3()
       : base(null, null, (long)600851475143) {
-      SolutionResponse = 6857;
+        SolutionResponse = (long)6857;
     }
 
     public override object Run(RunModes runMode, object input, bool Logging) {
