+++ Euler/Problems/EulerProblem5.cs
@@ -9,8 +9,8 @@ namespace Euler.Problems {
 
   public class EulerProblem5 : Problem {
     public EulerProblem5()
-      : base(10, 2520, 20) {
-      SolutionResponse = 232792560;
+      : base(10, (long)2520, 20) {
+        SolutionResponse = (long)232792560;
     }
 
     public override object Run(RunModes runMode, object input, bool Logging) {
